♻️ refactor : #8 메뉴 좋아요 관련 파일 이름 변경 및 위치 이동
