+++ src/main/java/com/sparta/greeypeople/like/menu/MenuLikes.java
@@ -1,4 +1,4 @@
-package com.sparta.greeypeople.like.entity;
+package com.sparta.greeypeople.like.menu;
 
 import com.sparta.greeypeople.menu.entity.Menu;
 import com.sparta.greeypeople.timestamp.TimeStamp;
