+++ src/main/java/com/sparta/greeypeople/like/menu/MenuLikesController.java
@@ -1,4 +1,4 @@
-package com.sparta.greeypeople.like.controller;
+package com.sparta.greeypeople.like.menu;
 
 import lombok.RequiredArgsConstructor;
 
@@ -12,12 +12,11 @@ import com.sparta.greeypeople.exception.BadRequestException;
 import com.sparta.greeypeople.exception.DataNotFoundException;
 import com.sparta.greeypeople.exception.ForbiddenException;
 import com.sparta.greeypeople.exception.ViolatedException;
-import com.sparta.greeypeople.like.service.MenuLikesService;
 
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/stores/{storeId}/menus")
-public class LikeController {
+public class MenuLikesController {
 
 	private final MenuLikesService menuLikesService;
