+++ src/main/java/com/sparta/greeypeople/like/menu/MenuLikesRepository.java
@@ -1,12 +1,10 @@
-package com.sparta.greeypeople.like.repository;
+package com.sparta.greeypeople.like.menu;
 
 import com.sparta.greeypeople.menu.entity.Menu;
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
-import com.sparta.greeypeople.like.entity.MenuLikes;
 
 @Repository
 public interface MenuLikesRepository extends JpaRepository<MenuLikes, Long> {
